requirements.txt
torch>=1.7.0
torchvision
opencv-python
numpy
streamlit
Pillow
pandas
matplotlib
seaborn
scikit-learn